Fix issue with hard error when using Flywheel
For some reason, when an invalid product form is submitted from a
WordPress page running on a Flywheel site, Flywheel is intercepting that
error and overriding our pretty error message and showing a giant hard
error that doesn't make any sense and it throws off the layout of the
page.

This change adds support for catching the 503 that Flywheel is returning
and continues with showing the pretty error message that we previously
only showed for 500 errors.