Remove all remaining select2 references
The only one that was left was in the review pop-up in the admin, so I
replaced that with the existing choices.js library used everywhere else.

These changes also include a bunch of cleanup work in how the actual
choices js/css files were being loaded, including consolidating
everything into a single place.